Stability_GradientDescent_SuperResolution/requirements_txt.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
matplotlib>=3.3.0
scipy>=1.5.0
pandas>=1.1.0
pyyaml>=5.3.0
tqdm>=4.50.0
tensorboard>=2.4.0
jupyter>=1.0.0
pytest>=6.0.0
seaborn>=0.11.0
Pillow>=8.0.0
scikit-image>=0.18.0